Implemented presenter to fetch weather forecast from api and recycler view adapter to display the forecast in a recycler view

+++ app/src/main/java/com/pixelart/weatherappchallenge/model/Main.kt
@@ -15,5 +15,5 @@ data class Main(
     val groundLevel: Float,
     val humidity: Int,
     @SerializedName("temp_kf")
-    val tempKF: Int
+    val tempKF: Float
 )
