+++ app/src/main/java/com/pixelart/weatherappchallenge/model/Wind.kt
@@ -2,5 +2,5 @@ package com.pixelart.weatherappchallenge.model
 
 data class Wind(
     val speed: Double,
-    val deg: Int
+    val deg: Double
 )
